Rename mocks folder
