+++ src/mocks/handlers/github-repos-details.ts
